Flask(Main)/imagetotext/requirements.txt
Flask
Pillow
pytesseract
sentence-transformers
transformers
sumy
torch
nltk
